Extract chart into a seperate layer
